feat: add lambda support
