+++ codegen/src/lambda.rs
@@ -1,0 +1,5 @@
+use proc_macro::TokenStream;
+
+pub(crate) fn r#impl(_attr: TokenStream, item: TokenStream) -> TokenStream {
+    todo!()
+}
